Rearrange SRIM analysis
