+++ irradiapy/srimpy/__init__.py
@@ -1,4 +1,4 @@
 """This subpackage provides a Python interface for SRIM for simulating collisional cascades."""
 
-from irradiapy.srimpy import debris, injected, pka
+from irradiapy.srimpy import analysis
 from irradiapy.srimpy.srimdb import SRIMDB
